+++ Mlib/Physics/Advance_Times/Movables/Pitch_Look_At_Node.cpp
@@ -88,7 +88,7 @@ void PitchLookAtNode::set_absolute_model_matrix(const TransformationMatrix<float
         if (std::isnan(aim.aim_offset)) {
             return;
         }
-        t = aim.time;
+        t = (float)aim.time;
         offset(1) = aim.aim_offset;
     }
     RigidBodyPulses rbp = followed_->rbi_.rbp_;
